integrating the backend with the digital cuda backend
