+++ src/qilisdk/extras/cuda_backend/cuda_backend.py
@@ -13,11 +13,23 @@
 # limitations under the License.
 from __future__ import annotations
 
-from typing import Callable, Type, TypeVar
+import warnings
+from enum import Enum
+from typing import Callable, ClassVar, Type, TypeVar
 
 import cudaq
+import cupy as cp
 import numpy as np
+from cudaq import State
+from cudaq.operator import OperatorSum, ScalarOperator
+from cudaq.operator import Schedule as cuda_schedule
+from cudaq.operator import evolve, spin
 
+from qilisdk.analog.analog_backend import AnalogBackend
+from qilisdk.analog.analog_result import AnalogResult
+from qilisdk.analog.hamiltonian import Hamiltonian, PauliI, PauliOperator, PauliX, PauliY, PauliZ
+from qilisdk.analog.quantum_objects import QuantumObject
+from qilisdk.analog.schedule import Schedule
 from qilisdk.digital import (
     RX,
     RY,
@@ -47,7 +59,7 @@ TBasicGate = TypeVar("TBasicGate", bound=BasicGate)
 BasicGateHandlersMapping = dict[Type[TBasicGate], Callable[[cudaq.Kernel, TBasicGate, cudaq.QuakeValue], None]]
 
 
-class CudaBackend(DigitalBackend):
+class CudaBackend(DigitalBackend, AnalogBackend):
     """
     Digital backend implementation using CUDA-based simulation.
 
@@ -56,6 +68,8 @@ class CudaBackend(DigitalBackend):
     tensor network, and matrix product state simulations. Gate operations in the circuit are
     mapped to CUDA operations via dedicated handler functions.
     """
+
+    _PAULI_MAP: ClassVar[dict] = {PauliZ: spin.z, PauliX: spin.x, PauliY: spin.y, PauliI: spin.i}
 
     def __init__(self, simulation_method: DigitalSimulationMethod = DigitalSimulationMethod.STATE_VECTOR) -> None:
         """
@@ -186,6 +200,108 @@ class CudaBackend(DigitalBackend):
         handler(target_kernel, gate.basic_gate, qubit)
         kernel.adjoint(target_kernel, target_qubit)
 
+    def _hamiltonian_to_cuda(self, hamiltonian: Hamiltonian) -> OperatorSum:
+        out = None
+        for offset, terms in hamiltonian:
+            if out is None:
+                out = offset * np.prod([self._PAULI_MAP[pauli.__class__](pauli.qubit) for pauli in terms])
+            else:
+                out += offset * np.prod([self._PAULI_MAP[pauli.__class__](pauli.qubit) for pauli in terms])
+        return out
+
+    def evolve(
+        self,
+        schedule: Schedule,
+        initial_state: QuantumObject,
+        observables: list[PauliOperator | Hamiltonian],
+        store_intermediate_results: bool = False,
+        **kwargs: dict,
+    ) -> AnalogResult:
+        """computes the time evolution under of an initial state under the given schedule.
+
+        Args:
+            schedule (Schedule): The evolution schedule of the system.
+            initial_state (QuantumObject): the initial state of the evolution.
+            observables (list[PauliOperator  |  Hamiltonian]): the list of observables to be measured at the end of the evolution.
+            store_intermediate_results (bool): A flag to store the intermediate results along the evolution.
+
+        Raises:
+            ValueError: if the observables provided are not an instance of the PauliOperator or a Hamiltonian class.
+
+        Returns:
+            AnalogResult: The results of the evolution.
+        """
+        warnings.warn(
+            f"The evolve method only works with the 'dynamics' simulation method. Ignoring user selected method '{self.simulation_method.value}'"
+        )
+
+        cudaq.set_target("dynamics")
+
+        cuda_ham = None
+        steps = np.linspace(0, schedule.T, int(schedule.T / schedule.dt))
+
+        def parameter_values(time_steps: np.ndarray) -> cuda_schedule:
+
+            def compute_value(param_name: str, step_idx: int) -> float:
+
+                return schedule.get_coefficient(time_steps[int(step_idx)], param_name)
+
+            return cuda_schedule(list(range(len(time_steps))), list(schedule.hamiltonians), compute_value)
+
+        cuda_sched = parameter_values(steps)
+
+        def get_schedule(key: str) -> Callable:
+            return lambda **args: args[key]
+
+        cuda_ham = sum(
+            ScalarOperator(get_schedule(key)) * self._hamiltonian_to_cuda(ham)
+            for key, ham in schedule.hamiltonians.items()
+        )
+
+        cuda_obs = []
+        for obs in observables:
+            if isinstance(obs, PauliOperator):
+                cuda_obs.append(self._PAULI_MAP[obs.__class__](obs.qubit))
+            elif isinstance(obs, Hamiltonian):
+                cuda_obs.append(self._hamiltonian_to_cuda(obs))
+            else:
+                raise ValueError(f"unsupported observable type of {obs.__class__}")
+
+        evolution_result = evolve(
+            hamiltonian=cuda_ham,
+            dimensions=dict.fromkeys(range(schedule.nqubits), 2),
+            schedule=cuda_sched,
+            initial_state=State.from_data(cp.array(initial_state.to_dm().dense, dtype=cp.complex128)),
+            observables=cuda_obs,
+            collapse_operators=[],
+            store_intermediate_results=store_intermediate_results,
+        )
+
+        self._apply_simulation_method()
+
+        return AnalogResult(
+            final_expected_values=np.array(
+                [exp_val.expectation() for exp_val in evolution_result.final_expectation_values()]
+            ),
+            expected_values=(
+                np.array(
+                    [[val.expectation() for val in exp_vals] for exp_vals in evolution_result.expectation_values()]
+                )
+                if evolution_result.expectation_values() is not None
+                else None
+            ),
+            final_state=(
+                QuantumObject(np.array(evolution_result.final_state())).dag()
+                if evolution_result.final_state() is not None
+                else None
+            ),
+            intermediate_states=(
+                [QuantumObject(np.array(state)).dag() for state in evolution_result.intermediate_states()]
+                if evolution_result.intermediate_states() is not None
+                else None
+            ),
+        )
+
     @staticmethod
     def _handle_M(kernel: cudaq.Kernel, gate: M, circuit: Circuit, qubits: cudaq.QuakeValue) -> None:
         """
@@ -264,4 +380,5 @@ class CudaBackend(DigitalBackend):
     @staticmethod
     def _handle_U3(kernel: cudaq.Kernel, gate: U3, qubit: cudaq.QuakeValue) -> None:
         """Handle an U3 gate operation."""
+        kernel.u3(theta=gate.theta, phi=gate.phi, delta=gate.gamma, target=qubit)
         kernel.u3(theta=gate.theta, phi=gate.phi, delta=gate.gamma, target=qubit)
